# Speed up `coded_bkw` computation