--- a/estimator/lwe_bkw.py
+++ b/estimator/lwe_bkw.py
@@ -3,6 +3,7 @@
 See :ref:`Coded-BKW for LWE` for what is available.
 """
 from sage.rings.all import QQ
+from sage.functions.all import log
 from sage.functions.error import erf
 from sage.functions.other import floor, ceil
 from sage.misc.cachefunc import cached_function