Avoids symbolic computation at all cost for faster computation, especially since on line 62 there's a `sum` which computes the sum of (I believe) hundreds of expressions. Benchmarks on my laptop:

Kyber512: 277.6s -> 12.25s
Kyber1024: Long time -> 17.40s
TFHE630: 539.4s -> 20.15s

Look at that, 20x speed up! I'll also benchmark other places of the code and see what other fixes I have.

Depends on #76 for the changed import list. Hopefully it helps (beginning to?) resolve #74 and #75.